+++ pypgx/cli/create_regions_bed.py
@@ -40,15 +40,30 @@ PyPGx."""
 be removed too)."""
     )
     parser.add_argument(
+        '--target-genes',
+        action='store_true',
+        help=
+"""Whether to only return target genes, excluding control genes and
+paralogs."""
+    )
+    parser.add_argument(
         '--sv-genes',
         action='store_true',
         help=
-"""Whether to only return genes with SV."""
+"""Whether to only return target genes whose at least one star allele is
+defined by structural variation"""
     )
-
+    parser.add_argument(
+        '--var-genes',
+        action='store_true',
+        help=
+"""Whether to only return target genes whose at least one star allele is
+defined by SNVs/indels."""
+    )
 def main(args):
     bf = utils.create_regions_bed(
         assembly=args.assembly, add_chr_prefix=args.add_chr_prefix,
-        merge=args.merge, sv_genes=args.sv_genes
+        merge=args.merge, target_genes=args.target_genes,
+        sv_genes=args.sv_genes, var_genes=args.var_genes
     )
     sys.stdout.write(bf.to_string())